test: Update useImageLoad test
