+++ packages/insight-viewer/src/hooks/useImageLoad.spec.ts
@@ -1,0 +1,40 @@
+/* eslint-disable no-shadow */
+import { loadImage } from './useImageLoad'
+import { DefaultProp } from '../Viewer/const'
+import { CornerstoneImage } from '../utils/cornerstoneHelper'
+import { CORNERSTONE_IMAGE_MOCK } from '../mocks/const'
+
+const IMAGE_ID =
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000000.dcm'
+const defaultParam = {
+  onError: DefaultProp.onError,
+  requestInterceptor: DefaultProp.requestInterceptor,
+  imageId: IMAGE_ID,
+}
+const cornerstoneImage = CORNERSTONE_IMAGE_MOCK as unknown as CornerstoneImage
+
+describe('useImageLoad:', () => {
+  it('with invalid image', () => {
+    const getImage = async () => {
+      throw new Error('request fails')
+    }
+
+    const action = () =>
+      loadImage({
+        ...defaultParam,
+        getImage,
+      })
+    expect(action()).rejects.toThrow('request fails')
+  })
+
+  it('with valid image', () => {
+    const getImage = async () => cornerstoneImage
+
+    loadImage({
+      ...defaultParam,
+      getImage,
+    }).then(res => {
+      expect(res).toMatchObject(CORNERSTONE_IMAGE_MOCK)
+    })
+  })
+})
